refactor: store cost as float8 in DB, remove frontend cents conversion

+++ main.py
@@ -10,7 +10,7 @@ app = FastAPI()
 
 class MenuItem(BaseModel):
     itemName: str
-    itemCost: int
+    itemCost: float
     itemType: str
